invoice_reader/requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
easyocr==1.7.2
filelock==3.17.0
fsspec==2025.2.0
gitdb==4.0.12
GitPython==3.1.44
groq==0.18.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.37.0
Jinja2==3.1.5
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain-core==0.3.40
langchain-groq==0.2.4
langsmith==0.3.11
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.28.0
networkx==3.4.2
ninja==1.11.1.3
numpy==2.2.3
opencv-python-headless==4.11.0.86
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
pillow==11.1.0
protobuf==5.29.3
pyarrow==19.0.1
pyclipper==1.3.0.post6
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
pypdf==5.3.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.23.1
scikit-image==0.25.2
scipy==1.15.2
setuptools==75.8.2
shapely==2.0.7
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.42.2
sympy==1.13.1
tenacity==9.0.0
tifffile==2025.2.18
toml==0.10.2
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
wheel==0.45.1
zstandard==0.23.0
